Consolidated transform publisher into stream.py, so that we have a single element handling all commands/streams
