+++ launch.sh
@@ -1,4 +1,3 @@
 #!/bin/bash
 
-python3 realsense/stream.py &
-python3 realsense/transform_publisher.py
+python3 realsense/stream.py
